+++ compose/project.py
@@ -7,8 +7,9 @@ from docker.errors import APIError
 
 from .config import get_service_name_from_net, ConfigurationError
 from .const import LABEL_PROJECT, LABEL_SERVICE, LABEL_ONE_OFF
-from .service import Service, check_for_legacy_containers
+from .service import Service
 from .container import Container
+from .legacy import check_for_legacy_containers
 
 log = logging.getLogger(__name__)
